Update twig extends
