+++ package.json
@@ -20,10 +20,14 @@
     "@angular/platform-browser-dynamic": "^14.2.0",
     "@angular/router": "^14.2.0",
     "@types/twig": "^1.12.2",
+    "libphonenumber-js": "^1.10.18",
     "lodash": "^4.17.15",
     "moment-timezone": "^0.5.27",
     "numeral": "^2.0.6",
+    "pluralize": "^8.0.0",
+    "qrcode": "^1.5.1",
     "rxjs": "~6.6.7",
+    "string-pixel-width": "^1.11.0",
     "tslib": "^2.0.0",
     "twig": "^1.13.3",
     "zone.js": "~0.11.4"
